--- DSA/DeveshSrivastava_590017127/Day2/answersDay2.cpp
@@ -1,188 +1,0 @@
-/**/
-/*Week 1 Day 2 (17/06/2025)*/
-/*1. Sum of All Even Numbers*/
-/**/
-/*A problem that combines conditional logic with accumulation to filter and sum
- * specific elements.*/
-/**/
-/*Given an array of integers nums of size n, calculate and return the sum of all
- * even numbers in the array.*/
-/**/
-/*This problem combines conditional logic with accumulation, teaching you to
- * filter elements based on specific criteria while performing mathematical
- * operations on the filtered data.*/
-/**/
-/*Your task: Calculate and return the sum of all even numbers in the array.*/
-/*Examples*/
-/**/
-/*Input:*/
-/**/
-/*nums = [1, 2, 3, 4, 5]*/
-/**/
-/*Output:*/
-/**/
-/*6*/
-/**/
-/*Input:*/
-/**/
-/*nums = [10, 15, 20]*/
-/**/
-/*Output:*/
-/**/
-/*30*/
-/**/
-/*Input:*/
-/**/
-/*nums = [1, 3, 5]*/
-/**/
-/*Output:*/
-/**/
-/*0*/
-/**/
-/*2. Find First and Last Element*/
-/**/
-/*A problem that teaches array indexing and accessing elements at specific
- * positions.*/
-/**/
-/*Given an array nums of size n, return the first element and the last element
- * of the array as a pair.*/
-/**/
-/*This problem teaches you about array indexing and how to access elements at
- * specific positions, which is fundamental for understanding array boundaries
- * and element positioning.*/
-/**/
-/*Your task: Return the first and last elements of the array.*/
-/*Examples*/
-/**/
-/*Input:*/
-/**/
-/*nums = [1, 2, 3, 4]*/
-/**/
-/*Output:*/
-/**/
-/*First: 1, Last: 4*/
-/**/
-/*Input:*/
-/**/
-/*nums = [7]*/
-/**/
-/*Output:*/
-/**/
-/*First: 7, Last: 7*/
-/**/
-/*Input:*/
-/**/
-/*nums = [5, 9, 2]*/
-/**/
-/*Output:*/
-/**/
-/*First: 5, Last: 2*/
-/**/
-/*3. Check if Two Arrays are Equal*/
-/**/
-/*A problem that introduces frequency counting and comparison techniques for
- * data structure equality.*/
-/**/
-/*Given two arrays, determine if both contain the same elements with the same
- * frequency, regardless of their order. Return true if they are equal, else
- * return false.*/
-/**/
-/*This problem introduces you to frequency counting and comparison techniques,
- * which are essential for understanding how to compare data structures and
- * verify equality beyond simple element-by-element comparison.*/
-/**/
-/*Your task: Determine if two arrays contain the same elements with the same
- * frequencies.*/
-/*Examples*/
-/**/
-/*Input:*/
-/**/
-/*nums1 = [1, 2, 3, 4], nums2 = [4, 3, 2, 1]*/
-/**/
-/*Output:*/
-/**/
-/*true*/
-/**/
-/*Input:*/
-/**/
-/*nums1 = [1, 2, 2, 3], nums2 = [1, 2, 3, 3]*/
-/**/
-/*Output:*/
-/**/
-/*false*/
-/**/
-/*Input:*/
-/**/
-/*nums1 = [1, 2, 3], nums2 = [1, 2, 3, 4]*/
-/**/
-/*Output:*/
-/**/
-/*false*/
-
-#include <iostream>
-#include <unordered_map>
-#include <vector>
-
-using namespace std;
-
-class Solution {
-
-  public:
-    int sumOfEvens(vector<int> &nums) {
-      int sum = 0;
-      for (int num : nums) {
-        if (num % 2 == 0) {
-          sum += num;
-        }
-      }
-      return sum;
-    }
-
-    pair<int, int> sumOfFirstLast(vector<int> &nums) {
-      if (nums.empty()) {
-        return {0, 0}; // Handle empty array case
-      }
-      return {nums.front(), nums.back()};
-    }
-
-    bool isArrayEqual(vector<int> &arr1, vector<int> &arr2) {
-      if (arr1.size() != arr2.size()) {
-        return false;
-      }
-      unordered_map<int, int> freq;
-      for (int num : arr1) {
-        freq[num]++;
-      }
-      for (int num : arr2) {
-        if (freq.find(num) == freq.end() || freq[num] == 0) {
-          return false;
-        }
-        freq[num]--;
-      }
-      return true;
-    }
-};
-
-int main() {
-
-  Solution solution;
-
-  vector<int> nums1 = {1, 2, 3, 4, 5, 6, 7, 8, 9, 10};
-
-  cout << "Sum of even numbers: " << solution.sumOfEvens(nums1) << endl;
-
-  vector<int> nums2 = {1, 4, 1, 12, 4, 5};
-
-  pair<int, int> firstLast = solution.sumOfFirstLast(nums2);
-
-  cout << "First element: " << firstLast.first
-       << ", Last element: " << firstLast.second << endl;
-  vector<int> arr1 = {1, 2, 3, 4};
-  vector<int> arr2 = {4, 3, 2, 1};
-  cout << "Are arrays equal? "
-       << (solution.isArrayEqual(arr1, arr2) ? "Yes" : "No") << endl;
-  vector<int> arr3 = {1, 2, 2, 3};
-  vector<int> arr4 = {1, 2, 3, 3};
-  cout << "Are arrays equal? "
-       << (solution.isArrayEqual(arr3, arr4) ? "Yes" : "No") << endl;
-}
